+++ src/MODULE.h
@@ -225,16 +225,16 @@ public:
     }
       
     MaybeFcolorMap fcolorMap_ = 
-      copy_prop<face_descriptor, std::string>(mesh, "f:color");
+        copy_prop<face_descriptor, std::string, EK>(mesh, "f:color");
     MaybeFcolorMap fcolorMap2_ = 
-      copy_prop<face_descriptor, std::string>(clipper, "f:color");
+      copy_prop<face_descriptor, std::string, EK>(clipper, "f:color");
     const bool hasColors = fcolorMap_.second && fcolorMap2_.second;
     MaybeFscalarMap fscalarMap_ = 
-      copy_prop<face_descriptor, double>(mesh, "f:scalar");
+      copy_prop<face_descriptor, double, EK>(mesh, "f:scalar");
     MaybeFscalarMap fscalarMap2_ = 
-      copy_prop<face_descriptor, double>(clipper, "f:scalar");
+      copy_prop<face_descriptor, double, EK>(clipper, "f:scalar");
     const bool hasScalars = fscalarMap_.second && fscalarMap2_.second;
-
+    
     std::size_t nfaces = mesh.number_of_faces();
     std::size_t nfaces_clipper = clipper.number_of_faces();
 
@@ -445,10 +445,10 @@ public:
     std::size_t nfaces = mesh.number_of_faces();
 
     MaybeFcolorMap fcolorMap_ = 
-      copy_prop<face_descriptor, std::string>(mesh, "f:color");
+      copy_prop<face_descriptor, std::string, EK>(mesh, "f:color");
     const bool hasColors = fcolorMap_.second;
     MaybeFscalarMap fscalarMap_ = 
-      copy_prop<face_descriptor, double>(mesh, "f:scalar");
+      copy_prop<face_descriptor, double, EK>(mesh, "f:scalar");
     const bool hasScalars = fscalarMap_.second;
 
     ClipVisitor2 vis;
@@ -621,10 +621,10 @@ public:
     std::size_t nfaces = mesh.number_of_faces();
 
     MaybeFcolorMap fcolorMap_ = 
-      copy_prop<face_descriptor, std::string>(mesh, "f:color");
+      copy_prop<face_descriptor, std::string, EK>(mesh, "f:color");
     const bool hasColors = fcolorMap_.second;
     MaybeFscalarMap fscalarMap_ = 
-      copy_prop<face_descriptor, double>(mesh, "f:scalar");
+      copy_prop<face_descriptor, double, EK>(mesh, "f:scalar");
     const bool hasScalars = fscalarMap_.second;
 
     ClipVisitor2 vis;
@@ -981,6 +981,7 @@ public:
 
 
   // ----------------------------------------------------------------------- //
+  // ----------------------------------------------------------------------- //
   Rcpp::XPtr<EMesh3> dual() {
     EMesh3 dualmesh = dualMesh(mesh);
     return Rcpp::XPtr<EMesh3>(new EMesh3(dualmesh), false);
@@ -988,11 +989,13 @@ public:
 
 
   // ----------------------------------------------------------------------- //
+  // ----------------------------------------------------------------------- //
   Rcpp::DataFrame edges() {
     return getEdges<EK, EMesh3, EPoint3>(mesh);
   }
 
 
+  // ----------------------------------------------------------------------- //
   // ----------------------------------------------------------------------- //
   Rcpp::IntegerVector facesAroundVertex(int v) {
     if(v >= mesh.number_of_vertices()) {
@@ -1010,6 +1013,7 @@ public:
   }
 
 
+  // ----------------------------------------------------------------------- //
   // ----------------------------------------------------------------------- //
   void fair(Rcpp::IntegerVector indices) {
     if(!CGAL::is_triangle_mesh(mesh)) {
@@ -1033,6 +1037,7 @@ public:
   }
 
 
+  // ----------------------------------------------------------------------- //
   // ----------------------------------------------------------------------- //
   Rcpp::XPtr<EMesh3> fillBoundaryHole(int border, bool fairhole) {
     std::vector<halfedge_descriptor> border_cycles;
@@ -1083,6 +1088,7 @@ public:
   }
 
 
+  // ----------------------------------------------------------------------- //
   // ----------------------------------------------------------------------- //
   Rcpp::List filterMesh(Rcpp::IntegerVector selectedFaces) {
 
@@ -1177,6 +1183,7 @@ public:
 
 
   // ----------------------------------------------------------------------- //
+  // ----------------------------------------------------------------------- //
   void fixManifoldness() {
     std::size_t nmv = PMP::duplicate_non_manifold_vertices(mesh);
     if(nmv > 0) {
@@ -1193,6 +1200,7 @@ public:
   }
 
 
+  // ----------------------------------------------------------------------- //
   // ----------------------------------------------------------------------- //
   Rcpp::NumericVector geoDists(const int index) {
     if(!CGAL::is_triangle_mesh(mesh)) {
@@ -1219,6 +1227,7 @@ public:
   }
 
 
+  // ----------------------------------------------------------------------- //
   // ----------------------------------------------------------------------- //
   Rcpp::List getBorders() {
     std::vector<halfedge_descriptor> border_cycles;
@@ -1255,24 +1264,30 @@ public:
 
 
   // ----------------------------------------------------------------------- //
+  // ----------------------------------------------------------------------- //
   Rcpp::NumericMatrix getFacesInfo() {
     if(!CGAL::is_triangle_mesh(mesh)) {
       Rcpp::stop("The mesh is not triangle.");
     }
-    Rcpp::CharacterVector rownames = {"cx", "cy", "cz", "area"};
-    Rcpp::NumericMatrix FacesInfo(4, mesh.number_of_faces());
+    Rcpp::CharacterVector rownames = 
+      {"cx", "cy", "cz", "ccx", "ccy", "ccz", "area"};
+    Rcpp::NumericMatrix FacesInfo(7, mesh.number_of_faces());
     int i = 0;
     for(face_descriptor fd : mesh.faces()) {
       auto vs = vertices_around_face(mesh.halfedge(fd), mesh).begin();
       EPoint3 p1 = mesh.point(*(vs++));
       EPoint3 p2 = mesh.point(*(vs++));
       EPoint3 p3 = mesh.point(*vs);
-      EPoint3 centroid = CGAL::centroid(p1, p2, p3);
-      EK::FT sarea = CGAL::squared_area(p1, p2, p3);
+      EPoint3 centroid     = CGAL::centroid(p1, p2, p3);
+      EPoint3 circumcenter = CGAL::circumcenter(p1, p2, p3);
+      EK::FT sarea         = CGAL::squared_area(p1, p2, p3);
       Rcpp::NumericVector col_i = {
         CGAL::to_double<EK::FT>(centroid.x()),
         CGAL::to_double<EK::FT>(centroid.y()),
         CGAL::to_double<EK::FT>(centroid.z()),
+        CGAL::to_double<EK::FT>(circumcenter.x()),
+        CGAL::to_double<EK::FT>(circumcenter.y()),
+        CGAL::to_double<EK::FT>(circumcenter.z()),
         sqrt(CGAL::to_double<EK::FT>(sarea))
       };
       FacesInfo(Rcpp::_, i++) = col_i;
@@ -1280,10 +1295,11 @@ public:
     Rcpp::rownames(FacesInfo) = rownames;
     return Rcpp::transpose(FacesInfo);
   }
+  
 
-
- // ----------------------------------------------------------------------- //
- Rcpp::List getFacesList() {
+  // ----------------------------------------------------------------------- //
+  // ----------------------------------------------------------------------- //
+  Rcpp::List getFacesList() {
     return getFaces<EMesh3>(mesh);
   }
 
@@ -2181,61 +2197,54 @@ public:
 
 
   // ----------------------------------------------------------------------- //
+  // ----------------------------------------------------------------------- //
   void writeFile(
-    Rcpp::String filename, const int precision, 
-    const bool binary, std::string comments,
-    Rcpp::Nullable<Rcpp::NumericMatrix> normals_,
-    Rcpp::Nullable<Rcpp::IntegerMatrix> fcolors_,
-    Rcpp::Nullable<Rcpp::IntegerMatrix> vcolors_
+      Rcpp::String filename, const int precision, 
+      const bool binary, std::string comments,
+      Rcpp::Nullable<Rcpp::NumericMatrix> normals_,
+      Rcpp::Nullable<Rcpp::IntegerMatrix> fcolors_,
+      Rcpp::Nullable<Rcpp::IntegerMatrix> vcolors_
   ) {
-    EMesh3 meshcopy;
-    MapBetweenFaces f2fmap_;
-    boost::associative_property_map<MapBetweenFaces> f2fmap(f2fmap_);
-    MapBetweenVertices v2vmap_;
-    boost::associative_property_map<MapBetweenVertices> v2vmap(v2vmap_);
-    CGAL::copy_face_graph(
-      mesh, meshcopy, 
-      CGAL::parameters::face_to_face_map(f2fmap).vertex_to_vertex_map(v2vmap)
-    );
+    Mesh3 meshcopy = epeck2epick(mesh);
     if(normals_.isNotNull()) {
       Rcpp::NumericMatrix normals(normals_);
-      CGALnormals_map vnormal = 
-        meshcopy.add_property_map<vertex_descriptor, EVector3>(
-          "v:normal", CGAL::NULL_VECTOR
+      Mesh3::Property_map<vxdescr, Vector3> vnormal = 
+        meshcopy.add_property_map<vxdescr, Vector3>(
+            "v:normal", CGAL::NULL_VECTOR
         ).first;
-      for(const auto& [vsource, vtarget] : v2vmap_) {
-        Rcpp::NumericVector normal = normals(Rcpp::_, int(vsource));
+      for(Mesh3::Vertex_index vd : meshcopy.vertices()) {
+        Rcpp::NumericVector normal = normals(Rcpp::_, int(vd));
         if(!Rcpp::NumericVector::is_na(normal(0))) {
-          vnormal[vtarget] = EVector3(normal(0), normal(1), normal(2));
+          vnormal[vd] = Vector3(normal(0), normal(1), normal(2));
         }
       }
     }
     if(fcolors_.isNotNull()) {
       Rcpp::IntegerMatrix fcolors(fcolors_);
-      EMesh3::Property_map<face_descriptor, Color> 
-        fcolor = meshcopy.add_property_map<face_descriptor, Color>(
+      Mesh3::Property_map<fdescr, Color> 
+        fcolor = meshcopy.add_property_map<fdescr, Color>(
           "f:color", Color()
         ).first;      
-      for(const auto& [fsource, ftarget] : f2fmap_) {
-        Rcpp::IntegerVector color = fcolors(Rcpp::_, int(fsource));
+      for(Mesh3::Face_index fd : meshcopy.faces()) {
+        Rcpp::IntegerVector color = fcolors(Rcpp::_, int(fd));
         unsigned char red   = color(0);
         unsigned char green = color(1);
         unsigned char blue  = color(2);
-        fcolor[ftarget] = Color(red, green, blue);
+        fcolor[fd] = Color(red, green, blue);
       }
     }
     if(vcolors_.isNotNull()) {
       Rcpp::IntegerMatrix vcolors(vcolors_);
-      EMesh3::Property_map<vertex_descriptor, Color> 
-        vcolor = meshcopy.add_property_map<vertex_descriptor, Color>(
+      Mesh3::Property_map<vxdescr, Color> 
+        vcolor = meshcopy.add_property_map<vxdescr, Color>(
           "v:color", Color()
         ).first;      
-      for(const auto& [vsource, vtarget] : v2vmap_) {
-        Rcpp::IntegerVector color = vcolors(Rcpp::_, int(vsource));
+      for(Mesh3::Vertex_index vd : meshcopy.vertices()) {
+        Rcpp::IntegerVector color = vcolors(Rcpp::_, int(vd));
         unsigned char red   = color(0);
         unsigned char green = color(1);
         unsigned char blue  = color(2);
-        vcolor[vtarget] = Color(red, green, blue);
+        vcolor[vd] = Color(red, green, blue);
       }
     }
     writeMeshFile(filename, precision, binary, comments, meshcopy);